refactor: SettingsView components, reorganize
