+++ SwiftDiffusion/SettingsView/SettingsSections/CustomViews/BrowseFileRow.swift
@@ -1,0 +1,46 @@
+//
+//  BrowseFileRow.swift
+//  SwiftDiffusion
+//
+//  Created by Justin Bush on 2/12/24.
+//
+
+import SwiftUI
+
+struct BrowseFileRow: View {
+  var labelText: String?
+  var placeholderText: String
+  @Binding var textValue: String
+  var browseAction: () async -> String?
+  
+  var body: some View {
+    VStack(alignment: .leading) {
+      if let label = labelText {
+        Text(label)
+          .font(.system(size: 14, weight: .semibold, design: .default))
+          .underline()
+          .padding(.vertical, 2)
+          .padding(.horizontal, 14)
+      }
+      HStack {
+        TextField(placeholderText, text: $textValue)
+          .textFieldStyle(RoundedBorderTextFieldStyle())
+          .font(.system(size: 11, design: .monospaced))
+          .disabled(true)
+        Button("Browse...") {
+          Task {
+            if let path = await browseAction() {
+              textValue = path
+            }
+          }
+        }
+      }
+      .padding(.bottom, 14)
+    }
+  }
+}
+
+
+#Preview {
+  SettingsView(userSettings: UserSettings.shared, selectedTab: .files)
+}
