+++ SwiftDiffusion/SettingsView/SettingsSections/GeneralSection.swift
@@ -1,0 +1,20 @@
+//
+//  GeneralSection.swift
+//  SwiftDiffusion
+//
+//  Created by Justin Bush on 2/12/24.
+//
+
+import SwiftUI
+
+struct GeneralSection: View {
+  @ObservedObject var userSettings: UserSettings
+  
+  var body: some View {
+    Text("Coming soon")
+  }
+}
+
+#Preview {
+  SettingsView(userSettings: UserSettings.shared, selectedTab: .general)
+}
